% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_covariate_matrix.R
\name{generate_covariate_matrix}
\alias{generate_covariate_matrix}
\title{Generate simulated batch and covariates-}
\usage{
generate_covariate_matrix(
  n_subject,
  p_batch,
  n_covariates,
  batchwise_offset = NULL,
  correlation_strength = "none"
)
}
\arguments{
\item{n_subject}{Number of subjects}

\item{p_batch}{Proportion of subjects in batch 1 ()}

\item{n_covariates}{Number of biological covariates to preserve (p)}

\item{batchwise_offset}{Vector describing how much each covariate's mean should be offset in batch 1 compared to batch 0. All covariates are generated under normal distribution with variance of 1.}

\item{correlation_strength}{"none", "weak", "moderate", or "strong"}
}
\value{
n x (n_covariates+1) matrix of covariates and batch. Batch is the first column.
}
\description{
Generates an n x (n_covariates+1) matrix of covariates and batch. Allows for specification of how batch is correlated with covariates through batchwise_offset and how much covariates are correlated with each other.
}
\examples{
\dontrun{
generate_covariate_matrix(n_subject = 1000, p_batch = 0.6,
n_covariates = 3,
batchwise_offset = c(0, 2, 3),
correlation_strength = "moderate")
}
}
