% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covbat_simulator.R
\name{generate_covbat_simulations}
\alias{generate_covbat_simulations}
\title{Generate simulations under the CovBat model}
\usage{
generate_covbat_simulations(
  n_sims,
  covariates,
  effect_sizes,
  pca_effect_sizes,
  delta,
  pca_delta,
  cores = NULL,
  pca_cutoff = 0.95
)
}
\arguments{
\item{n_sims}{Number of simulated datasets to generate}

\item{covariates}{A matrix with the covariates to be used to adjust for the batch effects. The first column should be a binary variable indicating the batch (0 or 1).}

\item{effect_sizes}{A matrix with the effect sizes for each feature for batch and biological covariates.}

\item{pca_effect_sizes}{A matrix with the effect sizes for each feature for batch and biological covariates in the covariance structure.}

\item{delta}{A scalar value that determines the strength of the scale batch effect.}

\item{pca_delta}{A scalar value that determines the strength of the scale batch effect in the covariance.}

\item{cores}{Number of cores to use. Default is NULL which uses 1 core.}

\item{pca_cutoff}{A scalar parameter controlling the number of principal components to retain. The default value of 0.95 means that batch effects will only be simulated in the principal components explaining up to 95\% of the total variance.}
}
\value{
A list of matrices with the CovBat simulated data.
}
\description{
This function generates simulated data with batch effects assuming the CovBat model.
}
\examples{
\dontrun{
covariates = generate_covariate_matrix(n_subject = 1000, p_batch = 0.6,
n_covariates = 3,
batchwise_offset = c(0, 2, 3),
correlation_strength = "moderate")
effect_sizes = generate_effect_sizes(300, 3, c(3, 2, 2, -2), rep(.5, 4))
pca_effect_sizes = generate_effect_sizes(300, 3, c(1, -1, 0, 1), rep(.3, 4))
generate_covbat_simulations(covariates, effect_sizes, pca_residuals,
delta = 1.2)
}
}
