% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_effect_sizes.R
\name{generate_effect_sizes}
\alias{generate_effect_sizes}
\title{Generate effect sizes}
\usage{
generate_effect_sizes(
  n_features,
  n_covariates,
  effect_size_means,
  effect_size_variances
)
}
\arguments{
\item{n_features}{The number of features in the study}

\item{n_covariates}{The number of covariates in the study (i.e., the number of columns in the design matrix)}

\item{effect_size_means}{A vector of length n_features + 1 containing the means of the normal distributions from which the effect sizes will be drawn. The first element corresponds to the batch effect, and the remaining elements correspond to the effect sizes for the covariates.}

\item{effect_size_variances}{A vector of length n_features + 1 containing the variances of the normal distributions from which the effect sizes will be drawn. The first element corresponds to the batch effect, and the remaining elements correspond to the effect sizes for the covariates.}
}
\value{
A matrix of size (n_covariates + 1) x n_features, where the first row contains the intercepts and the remaining rows contain the effect sizes for the predictors.
}
\description{
This function generates a matrix of effect sizes.
}
\examples{
\dontrun{
generate_effect_sizes(300, 3, c(3, 2, 2, -2), rep(.5, 4))
}
}
