% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pca_residuals.R
\name{generate_pca_residuals}
\alias{generate_pca_residuals}
\title{Generate batch effects in the covariance structure}
\usage{
generate_pca_residuals(
  n_subjects,
  n_features,
  covariates,
  pca_effects,
  pca_delta,
  pca_cutoff = 0.95
)
}
\arguments{
\item{n_subjects}{The number of subjects in the study}

\item{n_features}{The number of features in the study}

\item{covariates}{A matrix of size n_subjects x (1 + n_covariates) containing the covariates used to generate the batch effects. The first column should be binary and represent the batch, while the remaining columns should be continuous and represent other covariates that should be controlled for.}

\item{pca_effects}{A matrix of size (1 + n_covariates) x n_features containing the batch effects and biological effects. Not all effects will be used, since batch effects are only corrected in the first few eigendimensions.}

\item{pca_delta}{A scalar parameter controlling the magnitude of the batch effects in the PCA residuals. A value of 1 means there is no scale batch effects.}

\item{pca_cutoff}{A scalar parameter controlling the number of principal components to retain. The default value of 0.95 means that batch effects will only be simulated in the principal components explaining up to 95\% of the total variance.}
}
\value{
A matrix of size n_subjects x n_features containing the residuals with the specified batch effects in PC space.
}
\description{
This function generates a matrix of PCA residuals with specified batch effects. These batch effects will be rotated to hide them in a covariance matrix.
}
\examples{
\dontrun{
pca_effects = generate_effect_sizes(300, 3, c(3, 2, 2, -2), rep(.5, 4))
generate_pca_residuals(n_subjects = 1000, n_features = 300, covariates,
pca_effects, pca_delta = 1.5, pca_cutoff = 0.95)
}
}
